Create an organization with a test member account
A random suffix is generated and appended to the test member account
because a member account is permanently closed only after 90 days.
Meaning, if I wanted to spin up a new test member account within that
period of time, I'd have to use a new account name. The random suffix
helps with this.

The aws_organizations_account.test resource sets depends_on on the
aws_organizations_organization.main resource because an organization
must exist for the member account to be created under. Otherwise, as
I've noticed, the tofu executable can fail on creating the member
account.