Fix tracking the OpenTofu deps as Terraform deps
This was done because, by default, Renovate can track the OpenTofu
dependencies (e.g. providers, the tofu binary) as if they were Terraform
dependencies. The matchDatasources solution using the OpenTofu registry
is what was recommended from Renovate's documentation for part of the
OpenTofu dependencies. As for the tofu binary itself, this is handled by
an already created regex custom manager. In addition, a rule was created
to prevent the binary from being picked up as the hashicorp/terraform
package too.

A groupSlug field was set to 'all' for most rules so any packages
matched would go under the same branch Renovate creates for all the
other packages (part of the group:all preset).